solved remarks after pull-2 (comments are removed, Pokemons are moved into separate file, option to flip a card forward/back is added
